Move AbstractSettingsFragment to a new file
Fixes #3121
